+++ docs/ui/components/Collapsible/Collapsible.test.tsx
@@ -1,28 +1,28 @@
-import { fireEvent, render } from '@testing-library/react';
+import { fireEvent, render, screen } from '@testing-library/react';
 
 import { Collapsible } from '.';
 
 describe(Collapsible, () => {
   it('hides content by default', () => {
-    const { getByText } = render(<Collapsible summary="Summary">Content</Collapsible>);
-    expect(getByText('Summary')).toBeVisible();
-    expect(getByText('Content')).not.toBeVisible();
+    render(<Collapsible summary="Summary">Content</Collapsible>);
+    expect(screen.getByText('Summary')).toBeVisible();
+    expect(screen.getByText('Content')).not.toBeVisible();
   });
 
   it('shows content when opened', () => {
-    const { getByText } = render(<Collapsible summary="Summary">Content</Collapsible>);
-    fireEvent.click(getByText('Summary'));
-    expect(getByText('Summary')).toBeVisible();
-    expect(getByText('Content')).toBeVisible();
+    render(<Collapsible summary="Summary">Content</Collapsible>);
+    fireEvent.click(screen.getByText('Summary'));
+    expect(screen.getByText('Summary')).toBeVisible();
+    expect(screen.getByText('Content')).toBeVisible();
   });
 
   it('shows content when rendered with open', () => {
-    const { getByText } = render(
+    render(
       <Collapsible summary="Summary" open>
         Content
       </Collapsible>
     );
-    expect(getByText('Summary')).toBeVisible();
-    expect(getByText('Content')).toBeVisible();
+    expect(screen.getByText('Summary')).toBeVisible();
+    expect(screen.getByText('Content')).toBeVisible();
   });
 });