+++ packages/@expo/cli/src/install/index.ts
@@ -28,6 +28,7 @@ export const expoInstall: Command = async (argv) => {
         `--fix       Automatically update any invalid package versions`,
         chalk`--npm       Use npm to install dependencies. {dim Default when package-lock.json exists}`,
         chalk`--yarn      Use Yarn to install dependencies. {dim Default when yarn.lock exists}`,
+        chalk`--pnpm      Use pnpm to install dependencies. {dim Default when pnpm-lock.yaml exists}`,
         `-h, --help  Usage info`,
       ].join('\n'),
       [